+++ src/main/java/org/jboss/ci/tracker/common/objects/BuildDto.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright (C) 2015 Jan Tymel
  *
  * This program is free software: you can redistribute it and/or modify
@@ -25,15 +25,15 @@ import java.util.List;
  */
 public class BuildDto implements Serializable {
 
-    private Long id;
+    private Integer id;
     private String name;
     private String url;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
